Исправлены ошибки в классах FileBackedTaskManagerTest и FileBackedTaskManager

+++ test/main/java/ru/kermort/praktikum/taskmanager/manager/FileBackedTaskManagerTest.java
@@ -3,6 +3,8 @@ package ru.kermort.praktikum.taskmanager.manager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import ru.kermort.praktikum.taskmanager.enums.TaskStatus;
+import ru.kermort.praktikum.taskmanager.enums.TaskType;
 import ru.kermort.praktikum.taskmanager.tasks.EpicTask;
 import ru.kermort.praktikum.taskmanager.tasks.SubTask;
 import ru.kermort.praktikum.taskmanager.tasks.Task;
@@ -12,6 +14,7 @@ import java.nio.file.Files;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class FileBackedTaskManagerTest extends InMemoryTaskManagerTest {
     FileBackedTaskManager fbtm;
@@ -29,48 +32,21 @@ public class FileBackedTaskManagerTest extends InMemoryTaskManagerTest {
 
     @AfterEach
     void taskManagerStateAndFileMustBeEquals() {
-        if (fbtm.getAllTasks().size() == 0 && fbtm.getAllEpicTasks().size() == 0 && fbtm.getAllSubTasks().size() == 0) {
-            return;
-        }
-
-        List<String> listFromTaskManager = new ArrayList<>();
-        listFromTaskManager.add("id,type,name,status,description,epic");
-        List<String> listFromFile = new ArrayList<>();
-        for (Task task: fbtm.getAllTasks()) {
-            listFromTaskManager.add(fbtm.toString(task).replace("\n", ""));
-        }
-        for (EpicTask epicTask: fbtm.getAllEpicTasks()) {
-            listFromTaskManager.add(fbtm.toString(epicTask).replace("\n", ""));
-        }
-
-        for (SubTask subTask: fbtm.getAllSubTasks()) {
-            listFromTaskManager.add(fbtm.toString(subTask).replace("\n", ""));
-        }
-        Collections.sort(listFromTaskManager);
-
         try {
-            String fileText = Files.readString(fbtm.getFile().toPath());
-            String[] arrayFromFile = fileText.split("\n");
-            Arrays.sort(arrayFromFile);
-            listFromFile = List.of(arrayFromFile);
-            Files.delete(fbtm.getFile().toPath());
+            if (Files.exists(fbtm.getFile().toPath())) {
+                Files.delete(fbtm.getFile().toPath());
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
-
-        assertEquals(listFromFile, listFromTaskManager, "состояние менеджера отличается от записанного в файл");
     }
 
     @Test
     void saveAndLoadEmptyFileTest() {
-        fbtm.deleteAllTasks();
         try {
+            fbtm.save();
             String fileText = Files.readString(fbtm.getFile().toPath());
-            assertEquals(fileText, "id,type,name,status,description,epic\n",
-                    "файл после сохранения пустого менеджера должен содержать только заголовок");
-            File file = fbtm.getFile();
-            BufferedWriter writer = new BufferedWriter(new FileWriter(file));
-            writer.write("");
+            assertTrue(fileText.isEmpty(), "после сохранения менеджера без задач файл должен быть пустой");
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -78,37 +54,186 @@ public class FileBackedTaskManagerTest extends InMemoryTaskManagerTest {
         fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
         assertEquals(fbtm.getAllTasks().size() + fbtm.getAllEpicTasks().size() + fbtm.getAllSubTasks().size(), 0,
                 "при загрузке из пустого файла в менеджере не должно быть задач");
-
     }
 
     @Test
-    void loadFromFileTest() {
-        BufferedWriter writer = null;
-        File file = null;
+    @Override
+    void addNewTask() {
+        super.addNewTask();
+        List<String> listFromFile = fileToList(fbtm.getFile());
+        List<String> listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+    }
+
+    @Test
+    @Override
+    void addNewEpicTask() {
+        super.addNewEpicTask();
+        List<String> listFromFile = fileToList(fbtm.getFile());
+        List<String> listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+    }
+
+    @Test
+    @Override
+    void addNewSubTask() {
+        super.addNewSubTask();
+        List<String> listFromFile = fileToList(fbtm.getFile());
+        List<String> listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+    }
+
+    @Test
+    @Override
+    void subTaskIdShouldBeRemovedFromEpicIfSubTaskRemoved() {
+        super.subTaskIdShouldBeRemovedFromEpicIfSubTaskRemoved();
+        List<String> listFromFile = fileToList(fbtm.getFile());
+        List<String> listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+    }
+
+    @Test
+    @Override
+    void changeStatusTest() {
+        Task t1 = new Task("Заголовок задачи 1", "Описание задачи 1");
+        int t1Id = fbtm.addTask(t1);
+        Task t2 = new Task("Заголовок задачи 2", "Описание задачи 2");
+        int t2Id = fbtm.addTask(t2);
+
+        EpicTask ep1 = new EpicTask("Заголовок эпика 1", "Описание эпика 1");
+        int ep1Id = fbtm.addEpicTask(ep1);
+        SubTask st11 = new SubTask("Заголовок подзадачи 1 эпика 1", "Описание подзадачи 1 эпика 1", ep1.getId());
+        int st11Id = fbtm.addSubTask(st11);
+
+        EpicTask ep2 = new EpicTask("Заголовок эпика 2", "Описание эпика 2");
+        int ep2Id = fbtm.addEpicTask(ep2);
+        SubTask st21 = new SubTask("Заголовок подзадачи 1 эпика 2", "Описание подзадачи 1 эпика 2", ep2.getId());
+        SubTask st22 = new SubTask("Заголовок подзадачи 2 эпика 2", "Описание подзадачи 2 эпика 2", ep2.getId());
+        int st21Id = fbtm.addSubTask(st21);
+        int st22Id = fbtm.addSubTask(st22);
+
+        List<String> listFromFile = fileToList(fbtm.getFile());
+        List<String> listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+
+        st11.setStatus(TaskStatus.DONE);
+        st21.setStatus(TaskStatus.DONE);
+        fbtm.updateSubTask(st11);
+        fbtm.updateSubTask(st21);
+
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager,
+                "после изменения статуса подзадачи состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+
+        fbtm.deleteTask(t2.getId());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager,
+                "после удаления задачи состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+
+        fbtm.deleteEpicTask(ep2.getId());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager,
+                "после удаления эпика состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+
+        fbtm.deleteAllSubTasks();
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager,
+                "после удаления всех подзадач состояние менеджера записано в файл неправильно");
+
+        fbtm = FileBackedTaskManager.loadFromFile(fbtm.getFile());
+        listFromFile = fileToList(fbtm.getFile());
+        listFromManager = managerToList(fbtm);
+        assertEquals(listFromFile, listFromManager, "состояние менеджера загружено неправильно");
+    }
+
+    private static List<String> fileToList(File file) {
+        List<String> listFromFile = new ArrayList<>();
         try {
-            file =  new File(File.createTempFile("test", "tmp").toString());
-            writer = new BufferedWriter(new FileWriter(file));
-            writer.write("id,type,name,status,description,epic\n" +
-                    "1,TASK,t1,NEW,d,\n" +
-                    "2,TASK,t2,NEW,d,\n" +
-                    "3,EPIC,ep1,NEW,d,\n" +
-                    "4,EPIC,ep2,NEW,d,\n" +
-                    "5,SUBTASK,st1,NEW,d,3\n" +
-                    "6,SUBTASK,st2,NEW,d,3\n");
+            String fileText = Files.readString(file.toPath());
+            String[] arrayFromFile = fileText.split("\n");
+            Arrays.sort(arrayFromFile);
+            listFromFile = List.of(arrayFromFile);
         } catch (IOException e) {
             e.printStackTrace();
-        } finally {
-            if (writer != null) {
-                try {
-                    writer.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
         }
-        fbtm = FileBackedTaskManager.loadFromFile(file);
-        assertEquals(fbtm.getAllTasks().size(), 2, "в менеджер загружено неверное количество задач");
-        assertEquals(fbtm.getAllEpicTasks().size(), 2, "в менеджер загружено неверное количество эпиков");
-        assertEquals(fbtm.getAllSubTasks().size(), 2, "в менеджер загружено неверное количество подзадач");
+        return listFromFile;
+    }
+
+    private static List<String> managerToList(FileBackedTaskManager fbtm) {
+        List<String> listFromTaskManager = new ArrayList<>();
+        if (fbtm.getAllTasks().size() + fbtm.getAllEpicTasks().size() + fbtm.getAllSubTasks().size() > 0) {
+            listFromTaskManager.add(FileBackedTaskManager.HEADER.replace("\n", ""));
+        }
+
+        for (Task task: fbtm.getAllTasks()) {
+            listFromTaskManager.add(taskToString(task).replace("\n", ""));
+        }
+        for (EpicTask epicTask: fbtm.getAllEpicTasks()) {
+            listFromTaskManager.add(taskToString(epicTask).replace("\n", ""));
+        }
+
+        for (SubTask subTask: fbtm.getAllSubTasks()) {
+            listFromTaskManager.add(taskToString(subTask).replace("\n", ""));
+        }
+        Collections.sort(listFromTaskManager);
+        return listFromTaskManager;
+    }
+
+    private static String taskToString(Task task) {
+        String parentTaskId = "";
+        if (task.getTaskType() == TaskType.SUBTASK) {
+            parentTaskId = String.valueOf(((SubTask) task).getParentTaskId());
+        }
+
+        return task.getId() + "," +
+                task.getTaskType() + "," +
+                task.getTitle() + "," +
+                task.getStatus() + "," +
+                task.getDescription() + "," +
+                parentTaskId + "\n";
     }
 }
